Systray: don't force size upon resize requests.
The resize request events are now ignored. Some clients (deadbeef, pidgin) refuse to work
well with sizes different to the one requested. We try to follow the size hints whenever
possible, but if the client try to set an specific size we let it. If the icon gets too
large it will be naturally "clipped" by bh and the next icon to the right (which is raised
after the current one). Neither stalonetray nor tint2 nor awesome handle resize requests.